+++ Kernel/Time/HPET.cpp
@@ -7,7 +7,6 @@
 #include <AK/StringView.h>
 #include <Kernel/Debug.h>
 #include <Kernel/Firmware/ACPI/Parser.h>
-#include <Kernel/Interrupts/InterruptManagement.h>
 #include <Kernel/Memory/MemoryManager.h>
 #include <Kernel/Memory/TypedMapping.h>
 #include <Kernel/Sections.h>
